server/Trend_it/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
future==1.0.0
gunicorn==23.0.0
joblib==1.4.2
llvmlite==0.43.0
numba==0.60.0
numpy==2.0.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.5.2
scipy==1.14.1
serializers==0.2.4
six==1.16.0
sqlparse==0.5.2
threadpoolctl==3.5.0
tslearn==0.6.3
tzdata==2024.2
